Touching up the PHP docs, moved reduce station pairings to A_Station class
